requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
bcrypt==4.1.2
certifi==2023.11.17
deprecation==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-location-field==2.7.2
django-redis==5.4.0
django-rest-framework==0.1.0
django-silk==5.0.4
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-spectacular==0.26.5
exceptiongroup==1.2.0
flake8==6.1.0
gotrue==1.3.1
gprof2dot==2022.7.29
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
packaging==23.2
postgrest==0.11.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.2
redis==5.0.1
referencing==0.30.2
rpds-py==0.10.6
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
websockets==11.0.3
